Don't make synchronous methods async.
These are meant to emulate an async interface, but not to actually run
asynchronously. As such, they shouldn't use the "async" keyword, shouldn't
"await", and should instead just return a task object. Note that .NET 4.5
doesn't have the Task.CompletedTask static property, so we have to fake it
by making a task that is completed with a throwaway value.