Avoid deadlocks by not capturing the current context.
In ASP.net request handlers or UI code, await-ing a task without
ConfigureAwait(false) captures the current context for the task to resume
on. If the current context then blocks waiting for the task (e.g. uses
.Result), the task won't be able to resume and you'll deadlock. We don't
need the current http or UI context for RPC tasks, so opt-out of them.