Open core files with FILE_SHARE_DELETE on windows.
This allows core files and dumped images to be renamed or deleted while
sbcl is running, with the same unlink semantics as systems like Linux (i.e.
the file isn't actually deleted until the last reference to it is closed,
but the directory entry is removed). This requires us to use CreateFile()
instead of open(), which requires some trickery to work with perror()
reporting.