Remove bogus type expander.
An :INT parameter is passed as an immediate value and not as an allocated
object, so supplying a dynamic-extent allocation expansion doesn't do you
any good. Moreover, the previous code was creating the parameter as an int
pointer, which was the wrong type entirely. The default dynamic expansion
does the right thing in this case, so just use that.