Add from_param method to use class as type.
In order to use a class as a type, you have to provide a from_parm method
that will convert an object to the right internal type. In this case we're
just going to construct or return a BStr instance, since that already has
the appropriate conversion logic in it.